Updates all branches to 9.22
