+++ Cargo.toml
@@ -4,7 +4,6 @@ codegen-units = 1
 
 [workspace]
 members = [
-	"client",
 	"pallets/asset-registry",
 	"pallets/hasher",
 	"pallets/verifier",
@@ -20,10 +19,6 @@ members = [
 	"pallets/token-wrapper",
 	"pallets/token-wrapper-handler",
 	"standalone/*",
-	"open-runtime-module-library/currencies",
-	"open-runtime-module-library/tokens",
-	"open-runtime-module-library/traits",
-	"open-runtime-module-library/utilities",
 ]
 
 exclude = ["client"]
