--- common/Cargo.toml
@@ -1,64 +1,0 @@
-[package]
-name = "common"
-version = "1.0.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2018"
-description = "Logic which is common to all parachain runtimes"
-
-[package.metadata.docs.rs]
-targets = ['x86_64-unknown-linux-gnu']
-
-[dependencies]
-# External dependencies
-codec = { package = 'parity-scale-codec', version = '2.3.0', features = ['derive'], default-features = false }
-scale-info = { version = "1.0.0", default-features = false, features = ["derive"] }
-
-# Substrate dependencies
-sp-consensus-aura = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-sp-consensus-babe = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-sp-std = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-sp-io = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-frame-executive = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-frame-support = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-frame-system = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-pallet-balances = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-node-primitives = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-
-# Polkadot dependencies
-polkadot-runtime-common = { git = 'https://github.com/paritytech/polkadot', branch = "release-v0.9.17", default-features = false }
-polkadot-primitives = { git = 'https://github.com/paritytech/polkadot', branch = "release-v0.9.17", default-features = false }
-
-# Substrate dependencies
-pallet-treasury = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-
-# Webb primitives
-webb-primitives = { path = "../primitives", default-features = false }
-
-[dev-dependencies]
-sp-io = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-pallet-authorship = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-sp-keyring = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-
-[build-dependencies]
-substrate-wasm-builder = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17" }
-
-[features]
-default = ["std"]
-std = [
-	'codec/std',
-	'scale-info/std',
-	'sp-consensus-aura/std',
-	'sp-consensus-babe/std',
-	'sp-std/std',
-	'sp-io/std',
-	'frame-support/std',
-	'frame-executive/std',
-	'frame-system/std',
-	'pallet-treasury/std',
-	'pallet-balances/std',
-	'node-primitives/std',
-	'polkadot-runtime-common/std',
-	'polkadot-primitives/std',
-]
