--- common/src/lib.rs
@@ -1,98 +1,0 @@
-// Copyright (C) 2021 Parity Technologies (UK) Ltd.
-// SPDX-License-Identifier: Apache-2.0
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// 	http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#![cfg_attr(not(feature = "std"), no_std)]
-
-pub mod impls;
-pub use constants::*;
-pub use opaque::*;
-pub use types::*;
-
-/// Common types of parachains.
-mod types {
-	use sp_runtime::traits::{IdentifyAccount, Verify};
-
-	pub use webb_primitives::{ChainId, LeafIndex};
-
-	/// An index to a block.
-	pub type BlockNumber = u32;
-
-	/// Alias to 512-bit hash when used in the context of a transaction
-	/// signature on the chain.
-	pub type Signature = sp_runtime::MultiSignature;
-
-	/// Some way of identifying an account on the chain. We intentionally make
-	/// it equivalent to the public key of our transaction signing scheme.
-	pub type AccountId = <<Signature as Verify>::Signer as IdentifyAccount>::AccountId;
-
-	/// The type for looking up accounts. We don't expect more than 4 billion of
-	/// them, but you never know...
-	pub type AccountIndex = u32;
-
-	/// Balance of an account.
-	pub type Balance = u128;
-
-	/// Index of a transaction in the chain.
-	pub type Index = u32;
-
-	/// A hash of some data used by the chain.
-	pub type Hash = sp_core::H256;
-
-	/// Digest item type.
-	pub type DigestItem = sp_runtime::generic::DigestItem;
-
-	// Aura consensus authority.
-	pub type AuraId = sp_consensus_aura::sr25519::AuthorityId;
-
-	// Babe consensus authority.
-	pub type BabeId = sp_consensus_babe::AuthorityId;
-
-	/// Type used for expressing timestamp.
-	pub type Moment = u64;
-}
-
-/// Common constants of parachains.
-mod constants {
-	use frame_support::weights::{constants::WEIGHT_PER_SECOND, Weight};
-	use sp_runtime::Perbill;
-	/// We assume that ~5% of the block weight is consumed by `on_initialize`
-	/// handlers. This is used to limit the maximal weight of a single
-	/// extrinsic.
-	pub const AVERAGE_ON_INITIALIZE_RATIO: Perbill = Perbill::from_percent(5);
-	/// We allow `Normal` extrinsics to fill up the block up to 75%, the rest
-	/// can be used by Operational  extrinsics.
-	pub const NORMAL_DISPATCH_RATIO: Perbill = Perbill::from_percent(75);
-
-	/// We allow for 0.5 seconds of compute with a 6 second average block time.
-	pub const MAXIMUM_BLOCK_WEIGHT: Weight = WEIGHT_PER_SECOND / 2;
-}
-
-/// Opaque types. These are used by the CLI to instantiate machinery that don't
-/// need to know the specifics of the runtime. They can then be made to be
-/// agnostic over specific formats of data like extrinsics, allowing for them to
-/// continue syncing the network through upgrades to even the core data
-/// structures.
-pub mod opaque {
-	use super::*;
-	use sp_runtime::{generic, traits::BlakeTwo256};
-
-	pub use sp_runtime::OpaqueExtrinsic as UncheckedExtrinsic;
-	/// Opaque block header type.
-	pub type Header = generic::Header<BlockNumber, BlakeTwo256>;
-	/// Opaque block type.
-	pub type Block = generic::Block<Header, UncheckedExtrinsic>;
-	/// Opaque block identifier type.
-	pub type BlockId = generic::BlockId<Block>;
-}
