+++ pallets/mt/rpc/runtime-api/Cargo.toml
@@ -11,7 +11,7 @@ readme = "README.md"
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
-sp-api = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.17" }
+sp-api = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.22" }
 pallet-mt = { default-features = false, path = "../../" }
 webb-primitives = { path = "../../../../primitives", default-features = false }
