+++ primitives/src/types/mod.rs
@@ -1,4 +1,5 @@
 pub mod vanchor;
+pub mod runtime;
 
 use codec::{Decode, Encode};
 pub use ethabi::{encode, Token};
